+++ apps/doop/src/lib/store/createFiltersSlice.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { FILTER_TYPE_UNKNOWN } from "./createDataSlice"
+import { FILTER_TYPE_UNKNOWN } from "../filterViolations"
 import { valueToLabel } from "../helpers"
 
 // converts standardized filters object:
@@ -125,7 +125,7 @@ const createFiltersSlice = (set, get, options) => ({
             filters: { ...state.filters, active: [] },
           }),
           false,
-          "filters.remove"
+          "filters.remove_all"
         )
         // filter items
         get().data.actions.filterItems()
@@ -134,6 +134,7 @@ const createFiltersSlice = (set, get, options) => ({
         set((state: any) => ({
           filters: { ...state.filters, searchTerm },
         }))
+
         // filter items
         get().data.actions.filterItems()
       },