+++ apps/heureka/src/components/Services/Services.test.tsx
@@ -11,6 +11,7 @@ import { Services } from "./index"
 import { Filter, FilterSettings } from "../common/Filters/types"
 import { getTestRouter } from "../../mocks/getTestRouter"
 import { mockServicesPromise } from "../../mocks/promises"
+import { StoreProvider } from "../../store/StoreProvider"
 
 const mockFilters: Filter[] = [
   {
@@ -44,7 +45,9 @@ const renderComponent = () => {
     }),
     component: () => (
       <PortalProvider>
-        <Services />
+        <StoreProvider>
+          <Services />
+        </StoreProvider>
       </PortalProvider>
     ),
   })