+++ apps/carbon/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cloudoperators/juno-app-carbon
 
+## 1.0.19
+
+### Patch Changes
+
+- e53d025: Removed unused dependencies.
+
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [900b0be]
+  - @cloudoperators/juno-ui-components@5.6.0
+
 ## 1.0.17
 
 ### Patch Changes