+++ apps/greenhouse/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @cloudoperators/juno-app-greenhouse
 
+## 0.7.9
+
+### Patch Changes
+
+- e53d025: Removed unused dependencies.
+- Updated dependencies [e53d025]
+  - @cloudoperators/juno-url-state-provider@3.0.7
+  - @cloudoperators/juno-k8s-client@1.1.2
+  - @cloudoperators/juno-app-supernova@0.19.7
+  - @cloudoperators/juno-app-heureka@0.9.3
+  - @cloudoperators/juno-app-doop@2.6.3
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies [900b0be]
+  - @cloudoperators/juno-ui-components@5.6.0
+  - @cloudoperators/juno-app-doop@2.6.2
+  - @cloudoperators/juno-app-heureka@0.9.2
+  - @cloudoperators/juno-app-supernova@0.19.6
+  - @cloudoperators/juno-messages-provider@0.2.17
+
 ## 0.7.7
 
 ### Patch Changes