+++ apps/heureka/CHANGELOG.md
@@ -1,5 +1,21 @@
 # heureka
 
+## 0.9.3
+
+### Patch Changes
+
+- e53d025: Removed unused dependencies.
+- Updated dependencies [e53d025]
+  - @cloudoperators/juno-url-state-provider@3.0.7
+
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies [900b0be]
+  - @cloudoperators/juno-ui-components@5.6.0
+  - @cloudoperators/juno-messages-provider@0.2.17
+
 ## 0.9.1
 
 ### Patch Changes