+++ packages/url-state-provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cloudoperators/juno-url-state-provider
 
+## 3.0.7
+
+### Patch Changes
+
+- e53d025: Removed unused dependencies.
+
 ## 3.0.6
 
 ### Patch Changes